beanbot/requirements.txt
# Pip requirements file
#

# beanbot requirements
pyyaml
discord